updates pg connection pool and adds auto-repair; makes fronted more responsive
