+++ client/static/style.css
@@ -141,7 +141,7 @@ button {
 #no-data {
   position: absolute;
   top: 50%;
-  right: 40%;
+  left: 300px;
   align-items: center;
   text-align: center;
   justify-content: center;
@@ -732,6 +732,22 @@ pre {
 .apiLink a {
   color: rgb(35, 35, 255);
   text-emphasis-style: bold;
+}
+
+/* responsive design */
+@media screen and (max-width: 1500px) {
+  #cards-display {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media screen and (max-width: 1100px) {
+  #cards-display {
+    grid-template-columns: repeat(1, 1fr);
+  }
+  #no-data h1 {
+    font-size: 2rem;
+  }
 }
 
 /* phone sizing */
